tqqq_agent/requirements.txt
pandas
numpy
yfinance
lightgbm
torch
scikit-learn
joblib
python-dotenv
requests
